+++ SpicesUpdate@claudiux/files/SpicesUpdate@claudiux/cs/4.4/modules/cs_applets.py
@@ -2,7 +2,8 @@
 
 import sys
 from SUExtensionCore import ManageSpicesPage, DownloadSpicesPage
-from SettingsWidgets import SidePage, SettingsStack
+from SettingsWidgets import SidePage
+from xapp.SettingsWidgets import SettingsStack
 from SUSpices import SU_Spice_Harvester
 from gi.repository import GLib, Gtk, Gdk