+++ SpicesUpdate@claudiux/files/SpicesUpdate@claudiux/cs/4.4/modules/cs_desklets.py
@@ -2,7 +2,8 @@
 
 from SUExtensionCore import ManageSpicesPage, DownloadSpicesPage
 from SUSpices import SU_Spice_Harvester
-from GSettingsWidgets import *
+from SettingsWidgets import SidePage
+from xapp.GSettingsWidgets import *
 from gi.repository import GLib, Gtk
 
 class Module: